refactored into smaller functions
